packages refactored to 'ecb' pattern
