+++ javalab/src/main/java/com/smorales/javalab/middleware/config/JAXRSConfiguration.java
@@ -1,6 +1,6 @@
 package com.smorales.javalab.middleware.config;
 
-import com.smorales.javalab.middleware.buildtool.rest.WorkspaceProcessorResource;
+import com.smorales.javalab.middleware.workspaceprocessor.boundary.rest.WorkspaceProcessorResource;
 
 import javax.ws.rs.core.Application;
 import java.util.Set;
