+++ javalab/src/main/java/com/smorales/javalab/middleware/workspaceprocessor/boundary/NotRunnableCodeException.java
@@ -1,4 +1,4 @@
-package com.smorales.javalab.middleware.buildtool.boundary;
+package com.smorales.javalab.middleware.workspaceprocessor.boundary;
 
 public class NotRunnableCodeException extends RuntimeException {
