+++ javalab/src/main/java/com/smorales/javalab/middleware/workspaceprocessor/entity/Workspace.java
@@ -1,4 +1,4 @@
-package com.smorales.javalab.middleware.buildtool.entity;
+package com.smorales.javalab.middleware.workspaceprocessor.entity;
 
 import javax.persistence.*;
 import java.io.Serializable;
