+++ javalab/src/test/java/com/smorales/javalab/middleware/workspaceprocessor/boundary/BuildToolTest.java
@@ -1,4 +1,4 @@
-package com.smorales.javalab.middleware.buildtool.boundary;
+package com.smorales.javalab.middleware.workspaceprocessor.boundary;
 
 import org.junit.Before;
 import org.junit.Test;
